update private route var/link names
