+++ src/components/Helper.js
@@ -1,0 +1,7 @@
+export const authenticate = () => {
+ if (localStorage.getItem("user")) {
+   return JSON.parse(localStorage.getItem("currentUser"));
+ } else {
+   return false;
+ }
+};
